Add ScreenTransitionDashCount
