+++ Dialog/English.txt
@@ -274,6 +274,7 @@ MODOPTIONS_EXTENDEDVARIANTS_DISABLEREFILLSONSCREENTRANSITION=   Disable Refills 
 MODOPTIONS_EXTENDEDVARIANTS_RESTOREDASHESONRESPAWN= Restore Dashes on Respawn
 MODOPTIONS_EXTENDEDVARIANTS_RESTOREDASHESONRESPAWN_NOTE=    If you had 1 dash when you reached the spawn point, you will respawn with 1 dash
 MODOPTIONS_EXTENDEDVARIANTS_SPAWNDASHCOUNT=   Spawn dash count
+MODOPTIONS_EXTENDEDVARIANTS_SCREENTRANSITIONDASHCOUNT=   Screen transition dash count
 
 MODOPTIONS_EXTENDEDVARIANTS_DISPLAYDASHCOUNT=   Display Dash Count
