+++ Variants/DisableRefillsOnScreenTransition.cs
@@ -1,34 +1,15 @@
-﻿
-using System;
-using static ExtendedVariants.Module.ExtendedVariantsModule;
-
-namespace ExtendedVariants.Variants {
+﻿namespace ExtendedVariants.Variants {
     public class DisableRefillsOnScreenTransition : AbstractExtendedVariant {
-
+        /// <summary>
+        /// Prevents you from getting your dash and stamina back with screen transitions.
+        /// </summary>
+        /// <remarks>
+        /// See <see cref="ScreenTransitionDashCount.onPlayerTransition(On.Celeste.Player.orig_OnTransition, Celeste.Player)"/> for the code that sets the dash count on screen transition.
+        /// </remarks>
         public DisableRefillsOnScreenTransition() : base(variantType: typeof(bool), defaultVariantValue: false) { }
 
         public override object ConvertLegacyVariantValue(int value) {
             return value != 0;
-        }
-
-        public override void Load() {
-            On.Celeste.Player.OnTransition += onPlayerTransition;
-        }
-
-        public override void Unload() {
-            On.Celeste.Player.OnTransition -= onPlayerTransition;
-        }
-
-        private void onPlayerTransition(On.Celeste.Player.orig_OnTransition orig, Celeste.Player self) {
-            int bakDashes = self.Dashes;
-            float bakStamina = self.Stamina;
-
-            orig(self);
-
-            if (GetVariantValue<bool>(Variant.DisableRefillsOnScreenTransition)) {
-                self.Dashes = bakDashes;
-                self.Stamina = bakStamina;
-            }
         }
     }
 }
